Protect contact routes with JWT authentication and authorization

+++ controllers/contactController.js
@@ -2,18 +2,17 @@ const asyncHandler = require("express-async-handler");
 const Contact = require("../models/contactModel.js");
 //@desc get all contacts
 //@route GET /api/contacts
-//@access public
+//@access private
 const getContacts = asyncHandler(async (req, res) => {
-  const contacts = await Contact.find();
+  const contacts = await Contact.find({ user_id: req.user.id });
 
   res.status(200).json(contacts);
 });
 
 //@desc Create new Contact
 //@route POST /api/contacts
-//@access public
+//@access private
 const createContact = asyncHandler(async (req, res) => {
-  console.log("The request body is", req.body);
   const { name, email, phone } = req.body;
   if (!name || !email || !phone) {
     res.status(400);
@@ -23,13 +22,14 @@ const createContact = asyncHandler(async (req, res) => {
     name,
     email,
     phone,
+    user_id: req.user.id,
   });
 
   res.status(201).json(contact);
 });
 //@desc Get Contact
 //@route GET /api/contacts/:id
-//@access public
+//@access private
 const getContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
   if (!contact) {
@@ -40,12 +40,16 @@ const getContact = asyncHandler(async (req, res) => {
 });
 //@desc Update Contact
 //@route PUT /api/contacts/:id
-//@access public
+//@access private
 const updateContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
   if (!contact) {
     res.status(404);
     throw new Error("Contact Not Found");
+  }
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error("User don't have permission to update other user contacts");
   }
   const updatedContact = await Contact.findByIdAndUpdate(
     req.params.id,
@@ -57,12 +61,16 @@ const updateContact = asyncHandler(async (req, res) => {
 
 //@desc Delete Contact
 //@route DELETE /api/contacts/:id
-//@access public
+//@access private
 const deleteContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
   if (!contact) {
     res.status(404);
     throw new Error("Contact Not Found");
+  }
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error("User don't have permission to update other user contacts");
   }
   await Contact.findByIdAndDelete(req.params.id);
   res.status(200).json({ message: "Contact deleted successfully" });
